+++ Data/SavedData.lua
@@ -5,6 +5,72 @@ local data = addon.Data;
 data.SavedData = {};
 local savedData = data.SavedData;
 
+function savedData.TabsOrderAddIfNotContains(id, addonDisplayName, tabDisplayName)
+    SavedData.TabKeys = SavedData.TabKeys or {};
+
+    SavedData.TabKeys[id] = addonDisplayName .. " - " .. tabDisplayName;
+    if addon.Options.Defaults then -- Pre options loaded
+        addon.Options.Defaults.profile.Tabs[id].Order = id;
+    else -- Post options loaded
+        addon.Options.db.Tabs[id].Order = id;
+    end
+
+    addon.Options.InjectOptionsTableAdd({
+        order = id, type = "select", width = 2,
+        name = "",
+        values = function() return savedData.TabsOrderGetActiveKeys(); end,
+        get = function()
+            savedData.TabsOrderGetActiveKeys(); -- Just to make sure the list is cleaned up
+            for i, tab in next, addon.Options.db.Tabs do
+               if tab.Order == id then
+                   return i;
+               end
+            end
+        end,
+        set = function (_, value)
+            savedData.TabsOrderGetActiveKeys(); -- Just to make sure the list is cleaned up
+            for i, tab in next, addon.Options.db.Tabs do
+                if tab.Order == id then
+                    addon.Options.db.Tabs[i].Order = addon.Options.db.Tabs[value].Order;
+                end
+             end
+            addon.Options.db.Tabs[value].Order = id;
+            addon.GUI.ShowHideTabs();
+        end
+    }, tostring(id), "args", "Layout", "args", "Tabs", "args", "Order");
+end
+
+local needsCleanup = true;
+function savedData.TabsOrderGetActiveKeys()
+    if not needsCleanup then
+        return SavedData.TabKeys;
+    end
+    -- local tabsOrderActiveKeys = {};
+    -- local tabsToRemove = {};
+    for i = #addon.Options.db.Tabs, 1, -1 do
+        local tab = addon.Options.db.Tabs[i];
+        if tab.AddonName == "Blizzard_AchievementUI" or IsAddOnLoaded(tab.AddonName) then
+            -- tinsert(tabsOrderActiveKeys, SavedData.TabKeys[id]);
+        else
+            tremove(addon.Options.db.Tabs, i);
+            tremove(SavedData.TabKeys, i);
+        end
+    end
+    -- for id, tab in next, addon.Options.db.Tabs do
+    --     if tab.AddonName == "Blizzard_AchievementUI" or IsAddOnLoaded(tab.AddonName) then
+    --         tinsert(tabsOrderActiveKeys, SavedData.TabKeys[id]);
+    --     else
+    --         tinsert(tabsToRemove, id);
+    --     end
+    -- end
+    -- for _, id in next, tabsToRemove do
+    --     tremove(addon.Options.db.Tabs, id);
+    -- end
+    -- SavedData.TabKeys = tabsOrderActiveKeys;
+    needsCleanup = nil;
+    return SavedData.TabKeys;
+end
+
 local LoadSolutions, Resolve;
 function savedData.Load()
     if SavedData == nil then
@@ -37,7 +103,7 @@ function savedData.Load()
 end
 
 local FixFeaturesTutorialProgress, FixElvUISkin, FixFilters, FixEventDetails, FixShowExcludedCategory, FixEventDetails2, FixCharacters, FixEventAlert;
-local FixMergeSmallCategoriesThresholdChanged, FixShowCurrentCharacterIcons;
+local FixMergeSmallCategoriesThresholdChanged, FixShowCurrentCharacterIcons, FixTabs;
 function LoadSolutions()
     local solutions = {
         FixFeaturesTutorialProgress, -- 1
@@ -50,6 +116,7 @@ function LoadSolutions()
         FixEventAlert, -- 8
         FixMergeSmallCategoriesThresholdChanged, -- 9
         FixShowCurrentCharacterIcons, -- 10
+        FixTabs, -- 11
     };
 
     return solutions;
@@ -189,4 +256,39 @@ function FixShowCurrentCharacterIcons(prevBuild, currBuild, prevVersion, currVer
     addon.Options.db.Tooltip.Achievements.ShowCurrentCharacterIcons = nil;
 
     diagnostics.Debug("Cleared ShowCurrentCharacterIcons from previous version");
+end
+
+function FixTabs(prevBuild, currBuild, prevVersion, currVersion)
+    if currVersion < "35.0" or addon.Options.db.Tabs == nil or SavedData.Fixes.FixTabs == true then
+        diagnostics.Debug("Tabs already ported from previous version");
+        return;
+    end
+
+    for _addonName, _ in next, addon.Options.db.Tabs do
+        for _tabName, _ in next, addon.Options.db.Tabs[_addonName] do
+            print(_addonName, _tabName, type(addon.Options.db.Tabs[_addonName][_tabName]));
+            if type(addon.Options.db.Tabs[_addonName][_tabName]) ~= "table" then
+                addon.Options.db.Tabs[_addonName][_tabName] = {
+                    Show = addon.Options.db.Tabs[_addonName][_tabName];
+                };
+            end
+        end
+    end
+
+    StaticPopupDialogs["KROWIAF_FIXTABS"] = {
+        text = addon.MetaData.Title .. "\n\n" .. addon.L["FixTabs"] .. "\n\n - " .. addon.MetaData.Author,
+        button1 = addon.L["Options"],
+        button2 = addon.L["Close"],
+        OnButton1 = function()
+            addon.Options.Open();
+        end,
+        timeout = 0,
+        whileDead = true,
+        hideOnEscape = true
+    };
+    StaticPopup_Show("KROWIAF_FIXTABS");
+
+    SavedData.Fixes.FixTabs = true;
+
+    diagnostics.Debug("Ported Tabs from previous version");
 end