+++ Localization/deDE.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-06-04 19-09-36 ]] --
+-- [[ Exported at 2022-06-06 22-59-11 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --