+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 ## 9.2.5.35.0 - 2022-xx-xx
 ### Added
 - Calendar view for earned achievements (only 1st 4 per day for now)
+- Select All and Deselect All buttons for Event Reminders
 
 ### Changed
 - Allied race achievements now have a faction tied to them
+- Split the Event Reminders Events tab into separate tabs per event type
 
 ### Removed
 - Torcanata because it's a tracking achievement